moved test_fix_index_mismatch_from_upstream_array() out of separate file and into test_check_estimator_encoders.py where there are other tests of multiple encoders
