+++ tests/test_encoding/test_check_estimator_encoders.py
@@ -1,4 +1,7 @@
 import pytest
+import pandas as pd
+import numpy as np
+from sklearn.impute import SimpleImputer
 from sklearn.utils.estimator_checks import check_estimator
 
 from feature_engine.encoding import (
@@ -29,24 +32,71 @@ _estimators = [
     PRatioEncoder(ignore_format=True),
 ]
 
+@pytest.mark.parametrize(
+    "Estimator",
+    [
+        CountFrequencyEncoder(ignore_format=True),
+        DecisionTreeEncoder(regression=False, ignore_format=True),
+        MeanEncoder(ignore_format=True),
+        OneHotEncoder(ignore_format=True),
+        OrdinalEncoder(ignore_format=True),
+        RareLabelEncoder(
+            tol=0.00000000001,
+            n_categories=100000000000,
+            replace_with=10,
+            ignore_format=True,
+        ),
+        WoEEncoder(ignore_format=True),
+        PRatioEncoder(ignore_format=True),
+    ],
+)
+def test_all_transformers(Estimator):
+    return check_estimator(Estimator)
 
-@pytest.mark.parametrize("estimator", _estimators)
-def test_check_estimator_from_sklearn(estimator):
-    return check_estimator(estimator)
 
+@pytest.mark.parametrize(
+    # Encoders that encode X as a function of y; this is what
+    # breaks down when X becomes an array and indexes don't accidentally match in final
+    # concantenation
+    "encoder",
+    [MeanEncoder(), WoEEncoder(), PRatioEncoder()],
+)
+def test_fix_index_mismatch_from_upstream_array(encoder):
+    """
+    Created 2022-02-19 to test fix to issue # 376
+    Code adapted from:
+    https://github.com/scikit-learn-contrib/category_encoders/issues/280
+    """
 
-_estimators = [
-    CountFrequencyEncoder(),
-    DecisionTreeEncoder(regression=False),
-    MeanEncoder(),
-    OneHotEncoder(),
-    OrdinalEncoder(),
-    RareLabelEncoder(),
-    WoEEncoder(),
-    PRatioEncoder(),
-]
+    # test dataframe; setup for a transfromation where
+    # coded version of 'x' will be a function of target 'y'
+    df: pd.DataFrame = pd.DataFrame(
+        {
+            "x": ["a", "a", "b", "b", "c", "c"],
+            "y": [1, 0, 1, 0, 1, 0],
+        }
+    )
+    # Key - "non-standard" index that is not the usual
+    # contiguous range starting a t 0
+    df.index = [101, 105, 42, 76, 88, 92]
 
+    # Set up for standard pipeline/training etc.
+    X: pd.DataFrame = df[["x"]]
+    y: pd.Series = df["y"]
 
-@pytest.mark.parametrize("estimator", _estimators)
-def test_check_estimator_from_feature_engine(estimator):
-    return check_feature_engine_estimator(estimator)
+    # Will serve as a no-op whose chief purpose is to turn the
+    # X into an np.ndarray
+    si = SimpleImputer(strategy="constant", fill_value="a")
+
+    # Sequence leading to issue:
+    # 1) X becomes an array
+    assert type(X) == pd.DataFrame
+    X_2: np.array = si.fit_transform(X)
+    assert type(X_2) == np.ndarray
+
+    # 2) Encoder encodes as function of X, y
+    df_result: pd.DataFrame = encoder.fit_transform(X_2, y)
+    assert type(df_result) == pd.DataFrame
+
+    # Assertion fails: breakdown in index matches causes results to be all nan
+    assert all(df_result.iloc[:, 0].notnull())
